+++ src/xenia/cpu/hir/hir_builder.h
@@ -83,20 +83,20 @@ class HIRBuilder {
   void Trap(uint16_t trap_code = 0);
   void TrapTrue(Value* cond, uint16_t trap_code = 0);
 
-  void Call(FunctionInfo* symbol_info, uint32_t call_flags = 0);
+  void Call(FunctionInfo* symbol_info, uint16_t call_flags = 0);
   void CallTrue(Value* cond, FunctionInfo* symbol_info,
-                uint32_t call_flags = 0);
-  void CallIndirect(Value* value, uint32_t call_flags = 0);
-  void CallIndirectTrue(Value* cond, Value* value, uint32_t call_flags = 0);
+                uint16_t call_flags = 0);
+  void CallIndirect(Value* value, uint16_t call_flags = 0);
+  void CallIndirectTrue(Value* cond, Value* value, uint16_t call_flags = 0);
   void CallExtern(FunctionInfo* symbol_info);
   void Return();
   void ReturnTrue(Value* cond);
   void SetReturnAddress(Value* value);
 
-  void Branch(Label* label, uint32_t branch_flags = 0);
-  void Branch(Block* block, uint32_t branch_flags = 0);
-  void BranchTrue(Value* cond, Label* label, uint32_t branch_flags = 0);
-  void BranchFalse(Value* cond, Label* label, uint32_t branch_flags = 0);
+  void Branch(Label* label, uint16_t branch_flags = 0);
+  void Branch(Block* block, uint16_t branch_flags = 0);
+  void BranchTrue(Value* cond, Label* label, uint16_t branch_flags = 0);
+  void BranchFalse(Value* cond, Label* label, uint16_t branch_flags = 0);
 
   // phi type_name, Block* b1, Value* v1, Block* b2, Value* v2, etc
